requirements.txt
sanic
Sanic-Cors
tortoise-orm
# tortoise-orm[asyncmy] # driver of mysql
